# examhall/openapi.yaml
...
openapi: 3.0.3
info:
  title: ELEVIX - 1.0 Fullstack App Contract
  version: 1.0.0
  description: >
    Locked API contract for PS F1, F3, F4, F5, F7 and F8. Implement every
    path. The core logic must be handwritten - no auto-CRUD framework may
    generate these routes for you.
servers:
  - url: http://localhost:{port}
    variables:
      port:
        default: "8080"

# ---------------------------------------------------------------------------
# Every app must answer this, whatever the PS.
# ---------------------------------------------------------------------------
paths:
...
  /api/config:
    get:
      summary: F7 extension - retrieve the saved seating configuration.
      operationId: f7_get_config
      security: []
      responses:
        "200":
          description: Current groups, halls, and allocation settings.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/F7Config"}
    post:
      summary: F7 extension - save student groups and hall details.
      operationId: f7_save_config
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/F7Config"}
      responses:
        "200":
          description: Configuration saved.
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: {type: boolean}
        "400": {$ref: "#/components/responses/BadRequest"}
  /allocate:
    post:
      summary: F7 - run the constraint-satisfaction allocator. No OR-tools allowed.
      operationId: f7_allocate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [exam_id]
              properties:
                exam_id: {type: string}
                rules:
                  type: object
                  properties:
                    same_dept_adjacent: {type: boolean, default: false}
                    max_per_row:        {type: integer, default: 2}
      responses:
        "200":
          description: Allocation result.
          content:
            application/json:
              schema:
                type: object
                required: [exam_id, allocations, unseated]
                properties:
                  exam_id:     {type: string}
                  allocations:
                    type: array
                    items:
                      type: object
                      required: [hall, seat, student]
                      properties:
                        hall:    {type: string}
                        seat:    {type: string}
                        student: {type: string}
                  unseated:
                    type: array
                    items: {type: string}
        "400": {$ref: "#/components/responses/BadRequest"}
  /api/allocate:
    post:
      summary: F7 extension - allocate the saved configuration for the web application.
      operationId: f7_allocate_saved_config
      responses:
        "200":
          description: Seating plan generated from the saved configuration.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/F7SeatingPlan"}
        "422":
          description: No valid arrangement can be generated.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
  /api/allocation:
    get:
      summary: F7 extension - retrieve the latest seating plan.
      operationId: f7_get_allocation
      security: []
      responses:
        "200":
          description: Latest seating plan.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/F7SeatingPlan"}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/reallocate:
    post:
      summary: F7 extension - mark a hall unavailable and automatically re-seat its students.
      operationId: f7_reallocate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hallNo]
              properties:
                hallNo: {type: string}
      responses:
        "200":
          description: New seating plan after removing the unavailable hall.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/F7SeatingPlan"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409":
          description: Hall is already unavailable.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "422":
          description: Remaining halls cannot fit a valid arrangement.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
  /api/search/{reg}:
    get:
      summary: F7 extension - find a student's current hall and seat.
      operationId: f7_find_student_seat
      security: []
      parameters:
        - {name: reg, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: Current seat assignment for the register number.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/F7SearchResult"}
        "404": {$ref: "#/components/responses/NotFound"}

# ---------------------------------------------------------------------------
# F8  Library Book Exchange Platform
# ---------------------------------------------------------------------------
...
components:
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
          example: {error: BAD_REQUEST}
    NotFound:
      description: No such resource.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
          example: {error: NOT_FOUND}
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:   {type: string, description: A short machine-readable code.}
        message: {type: string, description: A human-readable explanation.}
...
    F7Config:
      type: object
      properties:
        strict:      {type: boolean, default: true}
        maxPerRow:   {type: integer, minimum: 1}
        cohorts:
          type: array
          items:
            type: object
            required: [dept, year, subject, startReg, endReg]
            properties:
              dept:         {type: string}
              year:         {type: integer}
              subject:      {type: string}
              startReg:     {type: string}
              endReg:       {type: string}
              lateralRegs:  {type: array, items: {type: string}}
              transferRegs: {type: array, items: {type: string}}
        halls:
          type: array
          items:
            type: object
            required: [floor, hallNo, rows, cols]
            properties:
              floor:          {type: integer}
              hallNo:         {type: string}
              rows:           {type: integer, minimum: 1}
              cols:           {type: integer, minimum: 1}
              maxDepartments: {type: integer, minimum: 1}
              maxStudents:    {type: integer, minimum: 1, description: Must not exceed rows multiplied by cols; defaults to physical capacity}
              blocked:
                type: array
                items: {type: string, pattern: '^\d+-\d+$'}
        invigilators:
          type: array
          items: {type: string}
    F7SeatingPlan:
      type: object
      required: [halls, conflicts, seated, generatedAt]
      properties:
        halls:
          type: array
          items:
            type: object
            required: [hallNo, floor, rows, cols, maxDepartments, maxStudents, invigilator, seats, subjects]
            properties:
              hallNo:         {oneOf: [{type: string}, {type: integer}]}
              floor:          {type: integer}
              rows:           {type: integer}
              cols:           {type: integer}
              maxDepartments: {type: integer}
              maxStudents:    {type: integer}
              invigilator:    {type: string}
              seats:
                type: array
                items:
                  type: object
                  required: [row, col, seatNo, reg, studentType, dept, year, subject]
                  properties:
                    row:         {type: integer}
                    col:         {type: integer}
                    seatNo:      {type: integer}
                    reg:         {type: string}
                    studentType: {type: string, enum: [normal, lateral, transfer]}
                    dept:        {type: string}
                    year:        {type: integer}
                    subject:     {type: string}
              subjects:
                type: array
                items:
                  type: object
                  required: [subject, dept, year]
                  properties:
                    subject: {type: string}
                    dept:    {type: string}
                    year:    {type: integer}
        conflicts:  {type: integer}
        seated:     {type: integer}
        generatedAt: {type: string, format: date-time}
    F7SearchResult:
      type: object
      required: [reg, studentType, hallNo, floor, row, col, seatNo, dept, year, subject]
      properties:
        reg:         {type: string}
        studentType: {type: string, enum: [normal, lateral, transfer]}
        hallNo:      {oneOf: [{type: string}, {type: integer}]}
        floor:       {type: integer}
        row:         {type: integer}
        col:         {type: integer}
        seatNo:      {type: integer}
        dept:        {type: string}
        year:        {type: integer}
        subject:     {type: string}
...
security: []